Add interactive Python notebook to use model for code explanations
This commit introduces an interactive Python notebook which is designed to use the "kreimben/CodeMind" pre-trained model and associated classes and functions. The notebook demonstrates how to use the model to generate responses, as well as how to format and tokenize inputs for model consumption. It also showcases how to use the output of the model to understand model performance and final detokenized response.