Add BERT-Large fine-tuning notebook
This commit includes the implementation of a new Jupyter notebook file for the fine-tuning of the BERT-Large model. The notebook comprises the setup of the environment, loading of the data and tokenizer, defining formatting and tokenization functions, and the training process. An EarlyStoppingCallback is also specified to prevent overfitting, and evaluation metrics are defined.