Update code for BERT-Large and GEMMA-2b-it models fine-tuning
This commit updates the code for the fine-tuning process of the BERT-Large and GEMMA-2b-it models. Changes include the addition of execution timings, output results, modification of output directories and project names, alteration of some training parameters, and rearrangement of code for better readability. A model upload step is also provided for pushing the trained model to Hugging Face.