Update text formatting in notebook
The text formatting for the prompt text and execution times was updated in the 'demo.ipynb' and 'gemma-2b-it peft qlora.ipynb' notebooks. This involved changing the import statement for Markdown, the way 'res' is displayed, as well as correcting the backslash escape sequence. Also, the instruction text in the generate_prompt function has been modified.