Add fine-tuning for model 'gemma-1.1-7b-it' with wandb and peft-qlora
The fine-tuning setup for the model 'gemma-1.1-7b-it' with weigth and biases and peft-qlora has been added through a new Jupyter notebook file. The added code includes the creation of configurations for the model, loading the model and tokenizer, modification of the model for training, loading the dataset, and building the training and testing datasets.