Add pandas library and update dataset loading process
The pandas library has been added to the dependencies of this script, and the method for loading the dataset has been changed to allow loading from a CSV file. Changes also included updating specific fields and parameters to accommodate the dataset's new format, including the renaming of certain fields and adjusting file paths.