Update eval.ipynb for improved model testing
Significant changes have been made in the 'eval.ipynb' notebook to improve testing of the model. This includes the introduction of additional imports, reworking the evaluation setup and adding new functions. Now, the model is set-up, dataset are prepared, and prompts are inputted for the model to predict, which is then evaluated using Rouge methodology.