Update code for fine-tuning and evaluating model
Code in eval.ipynb has been updated to include the finer details of fine-tuning and evaluating a pre-trained model. Changes include using a new import statement to load the model, and dataset name changes. The script now calculates accuracy and F1 score using the Trainer class from the Transformers library, which is newly imported. The review of the predicted example present in the previous version has been commented out.