Update gemma notebook with updated evaluation and training
Changes have been made to the gemma notebook to update the training and evaluation steps. Training is now performed using a new SFTTrainer and an updated dataset. Adjustments were also made to the content of the question and some of its parameters. The execution time was also added for various cells. The model evaluation and prediction samples have been removed in this update.