Update execution runtimes and model weights in gemma jupyter notebook
The execution times have been updated to reflect actual runtime of the training sessions in this notebook version. The model weights have also been re-organized and the specific run data is updated as per the new execution. Further, added GenerationConfig for Temperature, Sampling and Top-K parameters in order to allow control over the text generating process. Now the fine-tuned model in the notebook will be more transparent regarding its latency and effectiveness.