Update fine-tuning logic and adjust running steps
Changes made to the fine-tuning logic in gemma-1.1-2b-it peft qlora notebook file, including an adjustment to the max running steps from 500 to 300. Also, the text for GEMMA_2B_IT_MODEL_PREFIX_TEXT was updated, as well as the reordering projection layers and correction of execution times for the training cells.